# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "internet-date API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create a user and issue auth cookies",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthSignupBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSignupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify credentials and issue auth cookies",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Clear auth cookies",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Return current user id",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/feed": {
      "get": {
        "tags": [
          "feed"
        ],
        "summary": "Homepage feed (posts + match suggestions)",
        "security": [],
        "parameters": [
          {
            "name": "cursorId",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts": {
      "post": {
        "tags": [
          "feed"
        ],
        "summary": "Create post",
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreateBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCreateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{postId}": {
      "patch": {
        "tags": [
          "feed"
        ],
        "summary": "Update post",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPatchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{postId}/save": {
      "post": {
        "tags": [
          "feed"
        ],
        "summary": "Save post",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Id": {
        "type": "string",
        "pattern": "^\\d+$"
      },
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "AuthSignupBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthSignupResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/Id"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "email"
        ]
      },
      "AuthLoginBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthLoginResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/Id"
          }
        },
        "required": [
          "userId"
        ]
      },
      "AuthMeResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/Id"
          }
        },
        "required": [
          "userId"
        ]
      },
      "Visibility": {
        "type": "string",
        "enum": [
          "PUBLIC",
          "PRIVATE"
        ]
      },
      "MediaType": {
        "type": "string",
        "enum": [
          "IMAGE",
          "VIDEO"
        ]
      },
...
      "Media": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "url": {
            "type": "string"
          },
          "thumbUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "type": [
              "number",
              "null"
            ]
          },
          "height": {
            "type": [
              "number",
              "null"
            ]
          },
          "durationSec": {
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "type",
          "url"
        ]
      },
...
      "PostMedia": {
        "type": "object",
        "properties": {
          "order": {
            "type": "number"
          },
          "media": {
            "$ref": "#/components/schemas/Media"
          }
        },
        "required": [
          "order",
          "media"
        ]
      },
      "FeedPost": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "text": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "$ref": "#/components/schemas/Id"
              },
              "profile": {
                "type": [
                  "object",
                  "null"
                ],
                "properties": {
                  "displayName": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            },
            "required": [
              "id"
            ]
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostMedia"
            }
          }
        },
        "required": [
          "id",
          "createdAt",
          "user",
          "media"
        ]
      },
      "FeedSuggestion": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/Id"
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "locationText": {
            "type": [
              "string",
              "null"
            ]
          },
          "intent": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "userId"
        ]
      },
      "FeedResponse": {
        "type": "object",
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedPost"
            }
          },
          "suggestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedSuggestion"
            }
          },
          "nextCursorId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Id"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "posts",
          "suggestions",
          "nextCursorId"
        ]
      },
      "PostCreateBody": {
        "type": "object",
        "properties": {
          "text": {
            "type": [
              "string",
              "null"
            ]
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          },
          "mediaIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Id"
            }
          }
        }
      },
      "PostCreateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt"
        ]
      },
      "PostPatchBody": {
        "type": "object",
        "properties": {
          "text": {
            "type": [
              "string",
              "null"
            ]
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        }
      },
      "PostPatchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "text": {
            "type": [
              "string",
              "null"
            ]
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "visibility",
          "updatedAt"
        ]
      },
...
    }
  }
}